# src/routes/swagger/swagger.yaml
swagger: '2.0'
info:
  description: API
  version: 1.0.0
  title: Inventarios API
  # put the contact info for your development or API team
  contact:
    email: jhosepchai@gmail.com

  license:
    name: GNU GPL 3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html

# tags are used for organizing operations
...
paths:
...
  /Usuarios/listar:
    get:
      tags:
      - Usuarios
      summary: Lista usuarios
      operationId: listarUsuarios
      description: |
        Al hacer la peticion sin parametros se devolvera una lista de objetos de usuarios
      produces:
      - application/json

      security:
      - api_key: []
      
      responses:
        200:
          description: Devuelve array de usuarios activos
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        400:
          description: bad input parameter
...
  /Usuarios/buscarUsuarios/{Usuario}:
    get:
      tags:
      - Usuarios
      summary: Buscar usuarios
      operationId: buscarUsuarios
      description: |
        Al hacer la peticion con parametros se devolvera una lista de objetos de usuarios
      produces:
      - application/json
      parameters:
      - name: Usuario
        in: path
        description: Nombre de Usuario de la sucursal a buscar
        required: true
        type: string

      security:
      - api_key: []
      
      responses:
        200:
          description: Devuelve array de usuarios activos
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        400:
          description: bad input parameter
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "access-token"
    in: "header"
definitions:
...
  User:
    type: object
    properties:
      id:
        type: object
      usuario:
        type: string
      name:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
      createdAt:
        type: string
      estado:
        type: boolean
      
    xml:
      name: User
...
host: localhost:3000
basePath: /api
schemes:
 - https
 - http